Redefine Fibonacci

+++ code-session/src/main/java/com/wag/fib/Fibonacci.java
@@ -3,11 +3,20 @@ package com.wag.fib;
 public class Fibonacci {
 
     /**
-     * What is wrong with following code?
+     * Find nth fibonacci number.
+     *
+     * Fibonacci sequence is defined as follows:
+     * F(0) = 0
+     * F(1) = 1
+     * F(n) = F(n-1) + F(n-2)
+     * Example of sequence: 0, 1, 1, 2, 3, 5, 8....
+     * Example F(6) = 8
+     *
+     * What is wrong with following implementation?
      *
      * Change it so all unit tests are passed.
      */
-    public static int fibonacci(int n) {
+    public static long fibonacci(long n) {
         if (n <= 1) {
             return n;
         }
